pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bookforge"
version = "0.1.0"
description = "Automated publishing pipeline: manuscripts to EPUB, DOCX, and PDF"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1.0",
    "anthropic>=0.25.0",
    "openai>=1.30.0",
    "pytesseract>=0.3.10",
    "pymupdf>=1.24.0",
    "pypandoc>=1.13",
    "python-docx>=1.1.0",
    "ebooklib>=0.18",
    "weasyprint>=62.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "tiktoken>=0.7.0",
    "Jinja2>=3.1.0",
    "typer>=0.12.0",
    "structlog>=24.0.0",
    "chardet>=5.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
bookforge = "bookforge.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["bookforge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
